backend/requirements.txt
passlib
sqlalchemy
fastapi
pyjwt
uvicorn